calls_p4_b.rs hands its IntOp lookup tables and width wrapping to calls_p4_b_tail.rs, a pure text move under the 800-line ceiling
