+++ crates/almide-mir/src/lower/calls_p4_b.rs
@@ -383,437 +383,4 @@ impl LowerCtx {
         Some(dst)
     }
 
-    /// Extracted from `Self::lower_scalar_binop_shortcircuit_or_int` (eighth-round split,
-    /// cog reduction): the final eager `IntBinOp` fallback (with the narrow
-    /// signed-division-overflow guard), verbatim (only reached when the operator is NOT
-    /// `and`/`or` over Bool).
-    /// Extracted from `Self::lower_scalar_binop_int_fallback` (ninth-round split, cog
-    /// reduction): the pure `BinOp` + operand-shape → `IntOp` lookup, verbatim (a static
-    /// value computation, no `&mut self` needed).
-    /// Disjoint `BinOp` case (the 5 arithmetic ops, unguarded) of
-    /// [`Self::scalar_binop_int_op`] below — split out (codopsy cc); every arm here
-    /// is TOTAL once matched (no internal failure after a guard commits), so unlike
-    /// the guard-then-possibly-fail routers elsewhere, chaining via `.or_else()` is
-    /// safe: neither half can partially match and fall through wrongly.
-    fn scalar_binop_int_arith_op(op: &almide_ir::BinOp, left_ty: &Ty) -> Option<crate::IntOp> {
-        use almide_ir::BinOp;
-        // The unsigned 64-bit lane (#872): a `UInt64` operand's i64 slot
-        // carries the u64 bit pattern — add/sub/mul wrap identically in
-        // two's complement, but division/remainder must interpret it
-        // unsigned (the signed op computed `u64::MAX / 2` as `-1 / 2 = 0`).
-        let u64_lane = matches!(left_ty, Ty::UInt64);
-        Some(match op {
-            BinOp::AddInt => crate::IntOp::Add,
-            BinOp::SubInt => crate::IntOp::Sub,
-            BinOp::MulInt => crate::IntOp::Mul,
-            BinOp::DivInt if u64_lane => crate::IntOp::DivU,
-            BinOp::ModInt if u64_lane => crate::IntOp::ModU,
-            BinOp::DivInt => crate::IntOp::Div,
-            BinOp::ModInt => crate::IntOp::Mod,
-            _ => return None,
-        })
-    }
-
-    // Ordering comparisons (the `if` condition) — INT or BOOL operands (Bool is an i64 0/1,
-    // and v0's bool Ord is false < true = 0 < 1, so the i64 compare is bit-exact). A Float
-    // compare uses the prim float floor above; String ordering is the cmp-call above. Gate on
-    // the operand type. Disjoint `BinOp` case, split out (codopsy cc) of
-    // `scalar_binop_int_cmp_op` below.
-    fn scalar_binop_int_ord_op(op: &almide_ir::BinOp, left_ty: &Ty) -> Option<crate::IntOp> {
-        use almide_ir::BinOp;
-        // `UInt64` ordering is unsigned (#872): the signed compare put the
-        // upper half of the domain below zero.
-        if matches!(left_ty, Ty::UInt64) {
-            return Some(match op {
-                BinOp::Lt => crate::IntOp::LtU,
-                BinOp::Lte => crate::IntOp::LeU,
-                BinOp::Gt => crate::IntOp::GtU,
-                BinOp::Gte => crate::IntOp::GeU,
-                _ => return None,
-            });
-        }
-        Some(match op {
-            BinOp::Lt if Self::int_ord_operand_ty(left_ty) => crate::IntOp::Lt,
-            BinOp::Lte if Self::int_ord_operand_ty(left_ty) => crate::IntOp::Le,
-            BinOp::Gt if Self::int_ord_operand_ty(left_ty) => crate::IntOp::Gt,
-            BinOp::Gte if Self::int_ord_operand_ty(left_ty) => crate::IntOp::Ge,
-            _ => return None,
-        })
-    }
-
-    // Equality — INT or BOOL operands. A `Bool` is an i64 0/1 (a Var loads its 0/1, a
-    // `LitBool` materializes `ConstInt 0/1` above), so the SAME `IntOp::Eq`/`Ne` render is
-    // bit-exact for `b == false` / `b1 != b2` as for `n == 0`. (Ordering on Bool is undefined
-    // in v0, so it is NOT admitted; a Float/String/compound `==` still needs a distinct op.)
-    // Disjoint `BinOp` case, split out (codopsy cc) of `scalar_binop_int_cmp_op` below.
-    fn scalar_binop_int_eq_op(op: &almide_ir::BinOp, left_ty: &Ty) -> Option<crate::IntOp> {
-        use almide_ir::BinOp;
-        Some(match op {
-            BinOp::Eq if Self::int_eq_operand_ty(left_ty) => crate::IntOp::Eq,
-            BinOp::Neq if Self::int_eq_operand_ty(left_ty) => crate::IntOp::Ne,
-            _ => return None,
-        })
-    }
-
-    /// The comparison-op case (ordering + equality, INT/BOOL-operand-gated) of
-    /// [`Self::scalar_binop_int_op`] below — a thin router over the two disjoint-guard
-    /// helpers above (disjoint `BinOp` patterns from the arithmetic half too).
-    fn scalar_binop_int_cmp_op(op: &almide_ir::BinOp, left_ty: &Ty) -> Option<crate::IntOp> {
-        Self::scalar_binop_int_ord_op(op, left_ty).or_else(|| Self::scalar_binop_int_eq_op(op, left_ty))
-    }
-
-    // (Logical `and`/`or` are SHORT-CIRCUITED via control flow above — they never reach this
-    // eager `IntBinOp` path. Native + interp evaluate the RHS lazily. Pow, Float, concat,
-    // non-Int/Bool compares: defer — neither half above matches, so `None` falls through.)
-    fn scalar_binop_int_op(op: &almide_ir::BinOp, left_ty: &Ty) -> Option<crate::IntOp> {
-        Self::scalar_binop_int_arith_op(op, left_ty)
-            .or_else(|| Self::scalar_binop_int_cmp_op(op, left_ty))
-    }
-
-    fn lower_scalar_binop_int_fallback(
-        &mut self,
-        op: &almide_ir::BinOp,
-        left: &IrExpr,
-        right: &IrExpr,
-    ) -> Option<ValueId> {
-        // Either operand may be the one carrying the declared `UInt64` — a
-        // literal operand records the checker's default `Int` (#872), so
-        // asking only the left one missed `u64::MAX / 2`.
-        let lane_ty = if matches!(right.ty, Ty::UInt64) { &right.ty } else { &left.ty };
-        let iop = Self::scalar_binop_int_op(op, lane_ty)?;
-        let a = self.lower_scalar_value(left)?;
-        let b = self.lower_scalar_value(right)?;
-        self.emit_narrow_div_overflow_guard(iop, &left.ty, a, b);
-        let dst = self.fresh_value();
-        self.ops.push(Op::IntBinOp { dst, op: iop, a, b });
-        Some(self.narrow_wrap(dst, lane_ty, iop))
-    }
-
-    /// Re-wrap an arithmetic result to its DECLARED narrow width (#889).
-    ///
-    /// The MIR carries every integer in one i64, so `Int8 127 + 1` computes
-    /// `128` in the lane; native emits real `i8` arithmetic and wraps to
-    /// `-128`, so the wasm leg printed a value OUTSIDE the type's range and
-    /// the two targets disagreed. Wrapping at the declared width is the
-    /// documented semantics ("narrowing wraps rather than trapping",
-    /// stdlib/int8.almd), so re-apply it here for the ops that can leave the
-    /// range: signed → shift left then ARITHMETIC shift right (sign-extends
-    /// the truncated value), unsigned → mask. Comparisons and Int/Int64
-    /// (already the carrier's own width) are untouched, and `/`/`%` cannot
-    /// leave the range once their operands are in it — the one exception,
-    /// `MIN / -1`, already aborts via `emit_narrow_div_overflow_guard`.
-    fn narrow_wrap(&mut self, v: ValueId, ty: &Ty, iop: crate::IntOp) -> ValueId {
-        if !matches!(
-            iop,
-            crate::IntOp::Add | crate::IntOp::Sub | crate::IntOp::Mul
-        ) {
-            return v;
-        }
-        self.wrap_to_declared_width(v, ty)
-    }
-
-    /// Wrap `v` to `ty`'s declared width — [`Self::narrow_wrap`] without the
-    /// which-operator gate, for a producer that is not an `IntBinOp`.
-    ///
-    /// The `^` OPERATOR needs it: on wasm it lowers to a `math.pow` CALL, so it
-    /// never passed through the `IntBinOp` path that wraps, and `Int32 999997 ^ 2`
-    /// printed the full i64 `999994000009` while native — which computes at the
-    /// base's own width — printed the wrapped `-733379959`. `*` and `+` at the same
-    /// type already agreed, so the divergence was the operator's, not the type's.
-    /// Wrapping once at the end is exactly wrapping at each step: two's-complement
-    /// multiplication is congruent mod 2^bits, so a product of wrapped factors and
-    /// the wrap of the full product are the same value.
-    pub(crate) fn wrap_to_declared_width(&mut self, v: ValueId, ty: &Ty) -> ValueId {
-        let (bits, signed) = match ty {
-            Ty::Int8 => (8u32, true),
-            Ty::Int16 => (16, true),
-            Ty::Int32 => (32, true),
-            Ty::UInt8 => (8, false),
-            Ty::UInt16 => (16, false),
-            Ty::UInt32 => (32, false),
-            _ => return v,
-        };
-        if signed {
-            let shift = self.fresh_value();
-            self.ops.push(Op::ConstInt { dst: shift, value: (64 - bits) as i64 });
-            let up = self.fresh_value();
-            self.ops.push(Op::IntBinOp { dst: up, op: crate::IntOp::Shl, a: v, b: shift });
-            let down = self.fresh_value();
-            self.ops.push(Op::IntBinOp { dst: down, op: crate::IntOp::Shr, a: up, b: shift });
-            down
-        } else {
-            let mask = self.fresh_value();
-            let m = if bits == 64 { -1i64 } else { ((1u64 << bits) - 1) as i64 };
-            self.ops.push(Op::ConstInt { dst: mask, value: m });
-            let out = self.fresh_value();
-            self.ops.push(Op::IntBinOp { dst: out, op: crate::IntOp::And, a: v, b: mask });
-            out
-        }
-    }
-
-    /// Extracted from `Self::lower_scalar_binop_int_fallback` (ninth-round split, cog
-    /// reduction): the narrow signed-division-overflow guard injection, verbatim.
-    fn emit_narrow_div_overflow_guard(&mut self, iop: crate::IntOp, left_ty: &Ty, a: ValueId, b: ValueId) {
-        // NARROW signed division overflow (`Int8` MIN ÷ -1 — int8_div_overflow):
-        // the operands live in the i64 model, so the preamble's checked helper
-        // only catches i64::MIN ÷ -1; the narrow MIN wraps silently (v0 aborts
-        // "Error: integer overflow" + exit 1). Inject the width guard as MIR ops:
-        // if (a == MIN_w) & (b == -1) → prim.die with the SAME message bytes.
-        if !matches!(iop, crate::IntOp::Div | crate::IntOp::Mod) {
-            return;
-        }
-        let min_w = match left_ty {
-            Ty::Int8 => Some(-128i64),
-            Ty::Int16 => Some(-32768i64),
-            Ty::Int32 => Some(-2147483648i64),
-            _ => None,
-        };
-        let Some(mw) = min_w else { return };
-        let minc = self.fresh_value();
-        self.ops.push(Op::ConstInt { dst: minc, value: mw });
-        let negc = self.fresh_value();
-        self.ops.push(Op::ConstInt { dst: negc, value: -1 });
-        let c1 = self.fresh_value();
-        self.ops.push(Op::IntBinOp { dst: c1, op: crate::IntOp::Eq, a, b: minc });
-        let c2 = self.fresh_value();
-        self.ops.push(Op::IntBinOp { dst: c2, op: crate::IntOp::Eq, a: b, b: negc });
-        let both = self.fresh_value();
-        self.ops.push(Op::IntBinOp { dst: both, op: crate::IntOp::And, a: c1, b: c2 });
-        let msg = self.fresh_value();
-        self.ops.push(Op::Alloc {
-            dst: msg,
-            repr: crate::Repr::Ptr { layout: crate::PLACEHOLDER_LAYOUT },
-            init: crate::Init::Str("Error: integer overflow\n".into()),
-        });
-        let mh = self.fresh_value();
-        self.ops.push(Op::Prim { kind: crate::PrimKind::Handle, dst: Some(mh), args: vec![msg] });
-        self.ops.push(Op::IfThen { cond: both, dst: None });
-        self.ops.push(Op::Prim { kind: crate::PrimKind::Die, dst: None, args: vec![mh] });
-        self.ops.push(Op::Else { val: None });
-        self.ops.push(Op::EndIf { val: None });
-        // the message block is dead on the non-abort path — release it
-        self.ops.push(Op::Drop { v: msg });
-    }
-
-    /// Lower a `prim.*` PRIMITIVE-FLOOR call to an [`Op::Prim`] — the v1 self-host
-    /// floor (raw memory + the fd_write host call), mapped by name, NOT a real
-    /// `CallFn`/runtime symbol. Each arg lowers to a ValueId via
-    /// [`Self::lower_scalar_value`] (a handle var / int literal / int-arith). Returns
-    /// the result `dst` (load / fd_write / handle) or `None` (a store is Unit).
-    pub(crate) fn lower_prim_call(
-        &mut self,
-        func: &str,
-        args: &[IrExpr],
-    ) -> Result<Option<ValueId>, LowerError> {
-        // Each `matches!` mirrors the exact `func == "…"` name set the corresponding
-        // group's own blocks test — a router by NAME only, no behavior of its own; the
-        // group function is never called for a name outside its set.
-        if matches!(
-            func,
-            "alloc_str" | "alloc_bytes" | "alloc_list" | "alloc_list_f64" | "alloc_set"
-                | "alloc_map" | "alloc_value" | "alloc_list_str" | "alloc_set_str"
-                | "alloc_map_str" | "alloc_map_skv" | "alloc_map_kv" | "store_str"
-        ) {
-            return self.lower_prim_call_alloc(func, args);
-        }
-        if matches!(
-            func,
-            "args_get_list" | "args_get_list_full" | "env_get" | "read_text_file"
-                | "read_bytes_file" | "read_dir" | "write_text_file" | "make_dir"
-                | "remove_all" | "path_filestat" | "path_filestat_nofollow" | "path_exists"
-                | "rename"
-        ) {
-            return self.lower_prim_call_fs_env(func, args);
-        }
-        if matches!(func, "ptr_to_int" | "int_to_ptr" | "read_line" | "read_n_bytes") {
-            return self.lower_prim_call_ptr_io(func, args);
-        }
-        self.lower_prim_call_generic(func, args)
-    }
-
-    /// Extracted from `Self::lower_prim_call` (eleventh-round split, cog reduction): the
-    /// `alloc_*`/`store_str` name group, verbatim (only ever called for a name in the
-    /// router's matching `matches!` set, so no "unrecognized name" fallthrough is needed).
-    fn lower_prim_call_alloc(
-        &mut self,
-        func: &str,
-        args: &[IrExpr],
-    ) -> Result<Option<ValueId>, LowerError> {
-        if matches!(func, "alloc_str" | "alloc_bytes" | "alloc_list" | "alloc_list_f64" | "alloc_set" | "alloc_map" | "alloc_value") {
-            return self.lower_prim_call_alloc_scalar(func, args);
-        }
-        self.lower_prim_call_alloc_str(func, args)
-    }
-
-    /// Extracted from `Self::lower_prim_call_alloc` (twelfth-round split, cog reduction):
-    /// the scalar-element alloc name group, verbatim (only ever called for a name in the
-    /// caller's matching `matches!` set).
-    fn lower_prim_call_alloc_scalar(
-        &mut self,
-        func: &str,
-        args: &[IrExpr],
-    ) -> Result<Option<ValueId>, LowerError> {
-        // `prim.alloc_str(byte_len)` allocates a runtime-sized OWNED String — an `Op::Alloc`
-        // (cert `i`, a fresh owned object), NOT a scalar prim. The caller fills its bytes
-        // via `prim.store8`; the result is moved out / dropped like any heap value.
-        // `prim.alloc_str(n)` / `prim.alloc_bytes(n)` BOTH allocate a runtime-sized OWNED byte
-        // block (`Init::DynStr`: rc=1, len set, data filled by store8) — physically identical;
-        // they differ only in the prim's DECLARED return type (String vs Bytes). A flat heap
-        // value (no nested ownership), moved out / dropped like any String.
-        if func == "alloc_str" || func == "alloc_bytes" {
-            let len_v = self.lower_scalar_value(&args[0]).ok_or_else(|| {
-                LowerError::Unsupported(format!("prim.{func} length is not a lowerable scalar"))
-            })?;
-            let dst = self.fresh_value();
-            self.ops.push(Op::Alloc {
-                dst,
-                repr: crate::Repr::Ptr { layout: crate::PLACEHOLDER_LAYOUT },
-                init: crate::Init::DynStr { len: len_v },
-            });
-            return Ok(Some(dst));
-        }
-        // `prim.alloc_list(n)` allocates a runtime-sized OWNED `List[Int]` of n i64 slots —
-        // an `Op::Alloc` (cert `i`), the list-building sibling of alloc_str. The caller
-        // fills it via `prim.store64`; moved out / dropped like any heap value.
-        if func == "alloc_list" || func == "alloc_list_f64" || func == "alloc_set" || func == "alloc_map" || func == "alloc_value" {
-            let len_v = self.lower_scalar_value(&args[0]).ok_or_else(|| {
-                LowerError::Unsupported("prim.alloc_list length is not a lowerable scalar".into())
-            })?;
-            let dst = self.fresh_value();
-            self.ops.push(Op::Alloc {
-                dst,
-                repr: crate::Repr::Ptr { layout: crate::PLACEHOLDER_LAYOUT },
-                init: crate::Init::DynList { len: len_v },
-            });
-            return Ok(Some(dst));
-        }
-        unreachable!("lower_prim_call_alloc_scalar called with a name outside its caller-matched set: {func}")
-    }
-
-    /// Extracted from `Self::lower_prim_call_alloc` (twelfth-round split, cog reduction):
-    /// the heap-element (nested-ownership) alloc + `store_str` name group, verbatim (only
-    /// ever called for a name outside `lower_prim_call_alloc_scalar`'s set).
-    fn lower_prim_call_alloc_str(
-        &mut self,
-        func: &str,
-        args: &[IrExpr],
-    ) -> Result<Option<ValueId>, LowerError> {
-        use crate::PrimKind;
-        // `prim.alloc_list_str(n)` allocates a runtime-sized OWNED `List[String]` (n slots,
-        // physically identical to alloc_list) — but the dst is tracked as a NESTED-OWNERSHIP
-        // list, so its scope-end drop is a recursive `DropListStr` (frees the owned element
-        // Strings) and `prim.store_str` Consumes each String moved into it (Machinery 2).
-        if func == "alloc_list_str" || func == "alloc_set_str" || func == "alloc_map_str" || func == "alloc_map_skv" || func == "alloc_map_kv" {
-            let len_v = self.lower_scalar_value(&args[0]).ok_or_else(|| {
-                LowerError::Unsupported("prim.alloc_list_str length is not a lowerable scalar".into())
-            })?;
-            let dst = self.fresh_value();
-            self.ops.push(Op::Alloc {
-                dst,
-                repr: crate::Repr::Ptr { layout: crate::PLACEHOLDER_LAYOUT },
-                init: crate::Init::DynListStr { len: len_v },
-            });
-            self.value_drops.entry(dst).or_default().flat_elems = true;
-            return Ok(Some(dst));
-        }
-        // `prim.store_str(list, byte_addr_of_slot, piece)` — store the String `piece`'s handle
-        // into the list slot at `byte_addr_of_slot` AND CONSUME the piece (its reference is
-        // MOVED into the list, which now owns it — cert `m`, removed from the scope drop set).
-        // The slot holds the i64-widened handle; `DropListStr` later i32.wrap's it to free.
-        if func == "store_str" {
-            let addr = self.lower_scalar_value(&args[0]).ok_or_else(|| {
-                LowerError::Unsupported("prim.store_str slot address is not a lowerable scalar".into())
-            })?;
-            // The piece must be a tracked heap var (so we can Consume it). Its handle value:
-            let piece = match &args[1].kind {
-                IrExprKind::Var { id } => self.value_for(*id)?,
-                _ => {
-                    return Err(LowerError::Unsupported(
-                        "prim.store_str piece must be a heap variable (to consume)".into(),
-                    ))
-                }
-            };
-            // The slot value is the piece's HANDLE (its address as an i64). Op::Prim Handle
-            // gives that; store it 8-wide at the slot, then Consume the piece (move-out).
-            let handle = self.fresh_value();
-            self.ops.push(Op::Prim { kind: PrimKind::Handle, dst: Some(handle), args: vec![piece] });
-            self.ops.push(Op::Prim { kind: PrimKind::Store { width: 8 }, dst: None, args: vec![addr, handle] });
-            self.ops.push(Op::Consume { v: piece });
-            self.live_heap_handles.retain(|h| *h != piece);
-            return Ok(None);
-        }
-        unreachable!("lower_prim_call_alloc called with a name outside its router-matched set: {func}")
-    }
-
-    /// Extracted from `Self::lower_prim_call` (eleventh-round split, cog reduction): the
-    /// WASI env/fs-floor name group, verbatim (only ever called for a name in the
-    /// router's matching `matches!` set, so no "unrecognized name" fallthrough is needed).
-    fn lower_prim_call_fs_env(
-        &mut self,
-        func: &str,
-        args: &[IrExpr],
-    ) -> Result<Option<ValueId>, LowerError> {
-        if matches!(func, "args_get_list" | "args_get_list_full" | "env_get") {
-            return self.lower_prim_call_env(func, args);
-        }
-        self.lower_prim_call_fs(func, args)
-    }
-
-    /// Extracted from `Self::lower_prim_call_fs_env` (twelfth-round split, cog
-    /// reduction): the CLI-args/environ name group, verbatim (only ever called for a
-    /// name in the caller's matching `matches!` set).
-    fn lower_prim_call_env(
-        &mut self,
-        func: &str,
-        args: &[IrExpr],
-    ) -> Result<Option<ValueId>, LowerError> {
-        use crate::PrimKind;
-        // `prim.args_get_list()` — the WASI args→`List[String]` floor (env.args). NO
-        // args; its dst is a FRESH OWNED `List[String]` of `argv[1..]` (a heap Ptr, like
-        // an Alloc), so it is registered like `alloc_list_str`: a NESTED-OWNERSHIP list
-        // whose scope-end drop is the recursive `DropListStr` (frees the owned element
-        // Strings) — a flat `Drop` would leak them. Carries Capability::CliArgs (counted
-        // in cap_witness). The render emits the WASI args_sizes_get/args_get sequence.
-        if func == "args_get_list" {
-            let dst = self.fresh_value();
-            self.ops.push(Op::Prim { kind: PrimKind::ArgsGetList, dst: Some(dst), args: vec![] });
-            self.value_drops.entry(dst).or_default().flat_elems = true;
-            return Ok(Some(dst));
-        }
-        // `prim.args_get_list_full()` — the argv[0]-INCLUSIVE twin (process.args =
-        // std::env::args()). Same fresh OWNED List[String] + DropListStr + CliArgs
-        // discipline; renders through the SAME parameterized $args_get_list bridge.
-        if func == "args_get_list_full" {
-            let dst = self.fresh_value();
-            self.ops.push(Op::Prim { kind: PrimKind::ArgsGetListFull, dst: Some(dst), args: vec![] });
-            self.value_drops.entry(dst).or_default().flat_elems = true;
-            return Ok(Some(dst));
-        }
-        // `prim.env_get(name)` — the WASI environ lookup floor (env.get). ONE BORROWED
-        // `String` arg (the variable name; the caller still owns it). Its dst is a FRESH
-        // OWNED `Option[String]` in the `materialize_opt_str_some` layout (0-slot none /
-        // 1-slot some owning the value String @12), registered in `heap_elem_lists` so
-        // the scope-end drop is the flat `DropListStr` (frees the payload String, if
-        // any, then the block). Carries Capability::CliArgs — the Env profile's
-        // canonical cap (counted in cap_witness exactly like ArgsGetList).
-        if func == "env_get" && args.len() == 1 {
-            let key = match self.lower_call_args(args)?.into_iter().next() {
-                Some(CallArg::Handle(v)) => v,
-                _ => {
-                    return Err(LowerError::Unsupported(
-                        "prim.env_get needs a borrowed String name".into(),
-                    ))
-                }
-            };
-            let dst = self.fresh_value();
-            self.ops.push(Op::Prim { kind: PrimKind::EnvGet, dst: Some(dst), args: vec![key] });
-            self.value_drops.entry(dst).or_default().flat_elems = true;
-            return Ok(Some(dst));
-        }
-        // `env_get` with a wrong arg count (the ONLY name in this group with an extra
-        // guard beyond the bare name test) falls all the way through the original
-        // single-match to the terminal "unknown primitive" wall — replicated verbatim
-        // here (NOT `unreachable!`: this guard genuinely can fail for a matched name).
-        Err(LowerError::Unsupported(format!("unknown primitive prim.{func}")))
-    }
 }
